Refactor to webpack
