+++ assets/js/admin-csv-import-validation.js
@@ -1,0 +1,1 @@
+(()=>{"use strict";var e;e=function(){var e;e=function(e){var t=e.closest(".wpbdp-setting-row"),r=t.querySelector(".wpbdp-setting-label");return{mainWrapper:t,labelWrapper:r,label:r.querySelector("label")}},["input[name=csv-file]","input[name=images-file]"].forEach((function(t){var r=document.querySelector(t);r.addEventListener("change",(function(){!function(t){var r=e(t);r.label.style.color="",r.mainWrapper.querySelectorAll(".wpbdp-csv-import-error").forEach((function(e){return e.remove()})),0===document.querySelectorAll(".wpbdp-csv-import-error").length&&document.querySelectorAll(".submit input[type=submit]").forEach((function(e){return e.disabled=!1}))}(r);var t=r.files[0];t&&t.size>window.wpbdp_admin_import.maxFileSize&&function(t){var r=e(t);r.label.style.color="red";var n=document.createElement("p");n.classList.add("wpbdp-csv-import-error"),n.textContent=window.wpbdp_admin_import.error_label,r.labelWrapper.appendChild(n),document.querySelectorAll(".submit input[type=submit]").forEach((function(e){return e.disabled=!0}))}(r)}))}))},"undefined"!=typeof document&&("complete"!==document.readyState&&"interactive"!==document.readyState?document.addEventListener("DOMContentLoaded",e):e())})();
