+++ assets/js/admin-csv-import-validation/index.js
@@ -1,0 +1,6 @@
+import domReady from '@wordpress/dom-ready';
+import initializeCsvImportValidation from './initializeCsvImportValidation';
+
+domReady(() => {
+    initializeCsvImportValidation();
+});
