+++ webpack.config.js
@@ -13,6 +13,8 @@ const config = {
 	},
 	entry: {
 		'onboarding-wizard': './assets/js/onboarding-wizard/index.js',
+		'admin-csv-import-validation':
+			'./assets/js/admin-csv-import-validation/index.js',
 	},
 	output: {
 		filename: '[name].js',
